refactor(claude): split oauth into focused modules + key refreshed cache by source
Addresses review items #1 and #2 (structure/ownership, not behavior).

#1 — `providers/claude/oauth.rs` had grown to ~1138 lines mixing usage fetching,
refresh HTTP, refreshed-token caching, credential persistence, JSON patching and
tests. Split into a directory module:
  - oauth/mod.rs             — ClaudeOAuthFetcher: usage fetching + orchestration
                               (fetch/fetch_usage/rate-limit/snapshot) + types.
  - oauth/refresh.rs         — the OAuth token refresh HTTP call + its response type.
  - oauth/credentials_store.rs — credential loading (env/file/keyring), persistence,
                               and the refreshed-credentials cache.
The public surface is unchanged: `pub use oauth::ClaudeOAuthFetcher` still resolves
(directory module), and ClaudeOAuthFetcher's external API (new/fetch/
fetch_with_access_token) is untouched.

#2 — the refreshed-credential cache was a single process-global
`Option<ClaudeOAuthCredentials>` shared across every credential source. Because
`load_credentials` tries env→file→keyring and env tokens have no `expires_at`, a
file-refreshed token could shadow a freshly-read environment token (the
`(Some(_), None) => true` freshness arm). The cache is now
`HashMap<CredentialSource, _>` keyed by an explicit
`CredentialSource { Environment, File(path), Keyring(account) }`; `load_credentials`
returns the source and `ensure_fresh_credentials` threads it through the cache
lookup/store. Regression test `env_source_not_shadowed_by_file_cache` covers it.

Behavior-preserving. cargo check clean; core `codexbar` 490 tests, tauri 280 tests,
frontend 127 tests all green.
